Use git ls-remote instead of clone if not testing push
If the check is not verifying push access, then we don't need to clone
the entire repository in order to check it is available.  It is enough
to perform an operation which checks for the existence of a HEAD commit
on the remote repository.  The git ls-remote command can do that, and
will return a WARNING state if the repository is available but no HEAD
commit exists.